+++ src/main/java/org/netway/dongnehankki/global/auth/CustomUserDetailsService.java
@@ -1,14 +1,11 @@
 package org.netway.dongnehankki.global.auth;
 
 import lombok.RequiredArgsConstructor;
-import org.netway.dongnehankki.user.imfrastructure.UserRepository;
-import org.springframework.security.core.userdetails.User;
+import org.netway.dongnehankki.user.infrastructure.UserRepository;
 import org.netway.dongnehankki.global.exception.user.UnregisteredUserException;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.core.userdetails.UserDetailsService;
 import org.springframework.stereotype.Service;
-
-import java.util.Collections;
 
 @Service
 @RequiredArgsConstructor
