+++ src/main/java/org/netway/dongnehankki/user/presentation/UserController.java
@@ -4,8 +4,8 @@ import lombok.RequiredArgsConstructor;
 import org.netway.dongnehankki.global.common.ApiResponse;
 import org.netway.dongnehankki.user.application.UserService;
 import org.netway.dongnehankki.user.application.dto.login.LoginRequest;
-import org.netway.dongnehankki.user.application.dto.singup.CustomerSingUpRequest;
-import org.netway.dongnehankki.user.application.dto.singup.OwnerSingUpRequest;
+import org.netway.dongnehankki.user.application.dto.singUp.CustomerSingUpRequest;
+import org.netway.dongnehankki.user.application.dto.singUp.OwnerSingUpRequest;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
