+++ src/test/java/org/netway/dongnehankki/user/application/UserServiceTest.java
@@ -13,12 +13,12 @@ import org.netway.dongnehankki.global.exception.user.UnregisteredUserException;
 import org.netway.dongnehankki.store.domain.Store;
 import org.netway.dongnehankki.store.imfrastructure.StoreRepository;
 import org.netway.dongnehankki.user.application.dto.login.LoginRequest;
-import org.netway.dongnehankki.user.application.dto.singup.CustomerSingUpRequest;
-import org.netway.dongnehankki.user.application.dto.singup.OwnerSingUpRequest;
+import org.netway.dongnehankki.user.application.dto.singUp.CustomerSingUpRequest;
+import org.netway.dongnehankki.user.application.dto.singUp.OwnerSingUpRequest;
 import org.netway.dongnehankki.user.domain.User;
 import org.netway.dongnehankki.user.fixture.CustomerUserFixture;
 import org.netway.dongnehankki.user.fixture.OwnerUserFixture;
-import org.netway.dongnehankki.user.imfrastructure.UserRepository;
+import org.netway.dongnehankki.user.infrastructure.UserRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.security.authentication.AuthenticationManager;
