+++ src/test/java/org/netway/dongnehankki/user/presentation/UserControllerTest.java
@@ -12,12 +12,11 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import org.junit.jupiter.api.Test;
 import org.netway.dongnehankki.global.exception.user.DuplicateUserNameException;
 import org.netway.dongnehankki.global.exception.user.UnregisteredUserException;
-import org.netway.dongnehankki.store.domain.Store;
 import org.netway.dongnehankki.user.application.UserService;
 import org.netway.dongnehankki.user.application.dto.response.UserResponse;
 import org.netway.dongnehankki.user.application.dto.login.LoginRequest;
-import org.netway.dongnehankki.user.application.dto.singup.CustomerSingUpRequest;
-import org.netway.dongnehankki.user.application.dto.singup.OwnerSingUpRequest;
+import org.netway.dongnehankki.user.application.dto.singUp.CustomerSingUpRequest;
+import org.netway.dongnehankki.user.application.dto.singUp.OwnerSingUpRequest;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
 import org.springframework.boot.test.context.SpringBootTest;
